Don't combine `stocksize`, `fundertype` & `classification` sources
See #110

The IDs and values for the `stocksize`, `fundertype`, and
`classification` fields were combined from data in the Sigel
registry (IDs) and DBS (values). This resulted in missing or
inconsistent data. Instead, map Sigel IDs to their values (if we
have Sigel data), or DBS values to IDs (if we have no Sigel data).